Improve disconnection handling
Wake the thread with the associated_cv in the :close handler, as we might close
immediately, say if SSLv2 is attempted and fails.

Also when we timeout, return nil rather than false, so if needed the user can
test for nil/false/true to see if association didn't happen, failed, or
suceeded.