PCP-108 reduce the scope of @params
Here we refactor our constructor a little to reduce the scope of the params
hash.  Initially we lazily copied it into @params for later reference, but that
makes keeping track of what each element may be used for later more complex.
Now we promote all the important properties to instance variables and remove
@params.