PCP-109 Add devtools to bring in yard and yardstick targets
Yardstick is a YARD coverage checker, which will be useful to ensure
we don't accidentally miss documenting our API with YARD as we extend it.

Devtools adds a collection of rake targets that configure yardstick via
config/yardstick.yml so we can make targetted exclusions of yardstick rules.

Disable the ExampleTag rule that asserts all `@api public` tagged
methods are by default expected to have a corresponding `@example`
tag.  That's overkill most of the time for this codebase where many of the
methods are simple accessors.