PCP-109 Add basic YARD to all modules
This is the most basic level of YARD we need to get to 100% coverage of our
symbols.

For completelness we add a lib/pcp.rb to define `module PCP` so we
can add a documentation comment to it.  This is currently not used
at all by the code beyond being a place to hold that documentation.

```
$ bundle exec rake yard
Files:           4
Modules:         2 (    0 undocumented)
Classes:         2 (    0 undocumented)
Constants:       4 (    0 undocumented)
Methods:        16 (    0 undocumented)
 100.00% documented
```