(maint) reject a second connection on the same client
As we close over the @associated instance method in the callbacks having more
than one set of callbacks in flight is going to lead to odd interactions
between the multiple sets of callbacks and the object state.

It's possibly worth refactoring this logic out into a PCP::Client::Connection
class which can hold the state of each connection with it's associated
WebSocket::Client connection, but for now simply raise if called a second time.